+++ src/main/java/org/websoso/WSSServer/validation/UserIdValidator.java
@@ -8,7 +8,7 @@ import jakarta.validation.ConstraintValidatorContext;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.websoso.WSSServer.exception.exception.CustomUserException;
-import org.websoso.WSSServer.repository.UserRepository;
+import org.websoso.WSSServer.user.repository.UserRepository;
 
 @Component
 @AllArgsConstructor
